% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.default.R
\name{find.default}
\alias{find.default}
\title{Find - default}
\usage{
\method{find}{default}(object, ck)
}
\arguments{
\item{object}{a GSTSM object}

\item{ck}{set of candidates}
}
\value{
Solid Ranged-Group(s) of all candidate sequences
}
\description{
Default method for find. Does nothing.
}
