% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.gstsm.R
\name{find.gstsm}
\alias{find.gstsm}
\title{Find - GSTSM implementation}
\usage{
\method{find}{gstsm}(object, ck)
}
\arguments{
\item{object}{a GSTSM object}

\item{ck}{set of candidates}
}
\value{
Solid Ranged-Group(s) of all candidate sequences
}
\description{
GSTSM implementationfor for find method. Does nothing.
The goal is to find the Ranged Groups information for a candidate c.
}
