% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_adjacency_matrix.R
\name{generate_adjacency_matrix}
\alias{generate_adjacency_matrix}
\title{Generate Adjacency Matrix}
\usage{
generate_adjacency_matrix(spatial_positions, sigma)
}
\arguments{
\item{spatial_positions}{set of spatial positions}

\item{sigma}{max distance between group points}
}
\value{
Adjacency Matrix
}
\description{
Helper function that generates an adjacency matrix.
}
