% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_candidates.gstsm.R
\name{generate_candidates.gstsm}
\alias{generate_candidates.gstsm}
\title{Generate Candidates - GSTSM implementation}
\usage{
\method{generate_candidates}{gstsm}(object, srg)
}
\arguments{
\item{object}{a GSTSM object}

\item{srg}{set of Solid Ranged Groups}
}
\value{
candidate sequences of size k + 1
}
\description{
The algorithm combines SRGs that have sequences of size k, received as
input, to generate candidates with sequences of size k + 1. Let x and y be
SRGs, the conditions for this to occur are: that we have an
intersection of candidates over the time range,
intersection over the set of spatial positions (x.g n y.g), and
a common subsequence: <x.s2, . . . , x.sk>=<y.s1, . . . , y.sk-1>.
}
