% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.default.R
\name{merge.default}
\alias{merge.default}
\title{Merge - default}
\usage{
\method{merge}{default}(object, ck)
}
\arguments{
\item{object}{a GSTSM object}

\item{ck}{set of candidates}
}
\value{
Solid Ranged-Group(s) of all candidate sequences
}
\description{
Default method for merge. Does nothing.
}
