% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.gstsm.R
\name{merge.gstsm}
\alias{merge.gstsm}
\title{Merge - GSTSM implementation}
\usage{
\method{merge}{gstsm}(object, ck)
}
\arguments{
\item{object}{a GSTSM object}

\item{ck}{set of candidates}
}
\value{
Solid Ranged-Group(s) of all candidate sequences
}
\description{
Merge - GSTSM implementation
}
