% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{merge}
\alias{merge}
\title{Merge - definition}
\usage{
merge(object, ck)
}
\arguments{
\item{object}{a GSTSM object}

\item{ck}{set of candidates}
}
\value{
Solid Ranged-Group(s) of all candidate sequences
}
\description{
S3 class definition for merge method.
}
