% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_kernel_ranged_groups.R
\name{merge_kernel_ranged_groups}
\alias{merge_kernel_ranged_groups}
\title{Merge Kernel Ranged Groups}
\usage{
merge_kernel_ranged_groups(c, gamma)
}
\arguments{
\item{c}{candidate}

\item{gamma}{minimum temporal frequency}
}
\value{
KRG
}
\description{
The goal is to merge KRGs.
Let q and u be two different KRGs from the same candidate sequence.
They can be merged into a group qu = q U u as long as they have an
intersection and qu has a frequency greater than or equal to the minimum
frequency defined by the user.
}
