% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_open_kernel_ranged_groups.R
\name{merge_open_kernel_ranged_groups}
\alias{merge_open_kernel_ranged_groups}
\title{Merge Kernel Ranged Groups}
\usage{
merge_open_kernel_ranged_groups(c, timestamp, gamma, beta, adjacency_matrix)
}
\arguments{
\item{c}{candidate.}

\item{timestamp}{current timestamp}

\item{gamma}{minimum temporal frequency}

\item{beta}{minimum group size}

\item{adjacency_matrix}{adjacency matrix}
}
\value{
Set of updated KRGs
}
\description{
The goal of is to stretch KRGs of the same candidate sequence.
Its possible if two KRGs have intersection in space and
the resulting KRG keeps its frequency equal to or greater than beta.
}
