% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mine.default.R
\name{mine.default}
\alias{mine.default}
\title{Mine - default}
\usage{
\method{mine}{default}(object)
}
\arguments{
\item{object}{a GSTSM object}
}
\value{
all Solid Ranged Group(s) found, of all sizes
}
\description{
Default method for mine. Does nothing.
}
