% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mine.gstsm.R
\name{mine.gstsm}
\alias{mine.gstsm}
\title{Mine - GSTSM implementation}
\usage{
\method{mine}{gstsm}(object)
}
\arguments{
\item{object}{a GSTSM object}
}
\value{
all Solid Ranged Group(s) found, of all sizes
}
\description{
Mine - GSTSM implementation
}
