% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mine.R
\name{mine}
\alias{mine}
\title{Mine - definition}
\usage{
mine(object)
}
\arguments{
\item{object}{a GSTSM object}
}
\value{
all Solid Ranged Group(s) found, of all sizes
}
\description{
S3 class definition for mine method.
}
