% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_groups.R
\name{split_groups}
\alias{split_groups}
\title{Split Groups}
\usage{
split_groups(pos, adjacency_matrix)
}
\arguments{
\item{pos}{sequence occurrence index}

\item{adjacency_matrix}{possible connection between positions}
}
\value{
new set based on candidate c found in d.
}
\description{
Helper function that splits groups.
}
