% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_and_close.R
\name{validate_and_close}
\alias{validate_and_close}
\title{Validate and Close}
\usage{
validate_and_close(c, gamma, beta)
}
\arguments{
\item{c}{candidate}

\item{gamma}{minimum temporal frequency}

\item{beta}{minimum group size}
}
\value{
validated Greedy-Ranged-Groups.
}
\description{
The function receives as input the set of RGs (RG) from a candidate and the
minimum size of a group (beta).
It starts defining a set of elements that will be removed from the
set of RGs, if it does not have the minimum group size.
}
