% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_kernel_ranged_groups.R
\name{validate_kernel_ranged_groups}
\alias{validate_kernel_ranged_groups}
\title{Validate Kernel Ranged Groups}
\usage{
validate_kernel_ranged_groups(c, timestamp, gamma, beta)
}
\arguments{
\item{c}{candidate}

\item{timestamp}{current timestamp}

\item{gamma}{minimum temporal frequency}

\item{beta}{minimum group size}
}
\value{
Validated Kernel-Ranged-Groups.
}
\description{
Its objective is to verify that the user thresholds were observed
in each RGs, checking if they can still be stretched by keeping
the frequency greater than or equal to the minimum gamma and if
the minimum group size beta occurs.
It takes as input a set of RGs RG of a candidate sequence, the timestamp
of the start of the current sliding window timestamp, the user-defined
thresholds gamma and beta.
}
