% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_kernel_ranged_group.R
\name{find_kernel_ranged_group}
\alias{find_kernel_ranged_group}
\title{Find Kernel Ranged Group}
\usage{
find_kernel_ranged_group(c, d, gamma, beta, adjacency_matrix)
}
\arguments{
\item{c}{candidate}

\item{d}{set of transactions}

\item{gamma}{minimum temporal frequency}

\item{beta}{minimum group size}

\item{adjacency_matrix}{adjacency matrix}
}
\value{
Kernel Ranged-Group(s) of c updated
}
\description{
The goal is to find the Kernel Ranged Group information for a candidate c.
}
