% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_candidates.default.R
\name{generate_candidates.default}
\alias{generate_candidates.default}
\title{Generate Candidates - default}
\usage{
\method{generate_candidates}{default}(object, srg)
}
\arguments{
\item{object}{a GSTSM object}

\item{srg}{set of Solid Ranged Groups}
}
\value{
candidate sequences of size k + 1
}
\description{
Default method for generate_candidates. Does nothing.
}
