% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_candidates.R
\name{generate_candidates}
\alias{generate_candidates}
\title{Generate Candidates - definition}
\usage{
generate_candidates(object, srg)
}
\arguments{
\item{object}{a GSTSM object}

\item{srg}{set of Solid Ranged Groups}
}
\value{
candidate sequences of size k + 1
}
\description{
S3 class definition for generate_candidates method.
}
